requirements.txt
bs4
discord-py
requests
APScheduler==3.0.0